requirements.txt
Pillow
svg.path